GestorFinanciero/ocr-service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pytesseract==0.3.10
pillow==9.5.0
python-multipart==0.0.6
redis==4.5.5
numpy==1.24.3
pandas==2.0.1
pydantic==1.10.7
httpx==0.24.0
spacy==3.5.2